requirements.txt
chromadb
dashscope
python-docx
python-dotenv
PyYAML
